Fixed bug bug with index out of bounds in process_dir. Also removed telegram notifcation. Replaced tar with deflate to compress files.

+++ src/main.rs
@@ -1,16 +1,13 @@
 // This file would not be possible without Nekear
-
 use chrono::{DateTime, Datelike, Duration, Local, Timelike};
 use clap::Parser;
+use zip::write::FileOptions;
 use std::{
     fs::{self, DirEntry},
     path::PathBuf,
     process::exit,
-    time::Instant,
+    time::Instant, io::Write,
 };
-use ureq;
-
-use tar::Builder as TarBuilder;
 
 #[derive(Parser, Debug)]
 #[clap(author, version, about, long_about = None)]
@@ -25,18 +22,6 @@ struct Args {
     /// Delete files older than <delete> days old
     #[clap(short, long)]
     delete: usize,
-
-    /// Proxy in format <user>:<password>@<domain or ip>:<port (optional)>
-    #[clap(short, long, required = false)]
-    proxy_url: Option<String>,
-
-    /// Token of the bot which is used to send notifications
-    #[clap(short, long = "token", required = false)]
-    telegram_token: Option<String>,
-
-    /// Id of the chat where to send a notification
-    #[clap(short, long, required = false)]
-    chat_id: Option<String>,
 }
 
 fn main() {
@@ -64,94 +49,7 @@ fn main() {
 
     let processed = process_dir(&args.directory, &archive_from, &delete_from);
 
-    send_telegram_notification(&args, now.elapsed(), processed, delete_from, archive_from);
-}
-
-fn send_telegram_notification(args: &Args, elapsed_time: std::time::Duration, amount: usize, from_date: DateTime<Local>, to_date: DateTime<Local>) {
-    if args.telegram_token.is_none() {
-        println!("No notificaion send since no telegram token provided.");
-        return;
-    }
-
-    if args.chat_id.is_none() {
-        println!("Cannot send message without chat id provided!");
-        return;
-    }
-
-    let ip = get_host_ip();
-
-    if ip.is_none() {
-        println!("No valid host IP found to send notifcation to Telegram.");
-        return;
-    }
-
-    let http = get_http_client(&args.proxy_url);
-    if let Err(e) = &http {
-        println!("Bad proxy configuration: {e}");
-        return;
-    }
-
-    let ip = ip.unwrap();
-    let token = args.telegram_token.clone().unwrap();
-    let chat_id = args.chat_id.clone().unwrap();
-    let http = http.unwrap();
-
-    // TODO: Add ip info
-    let text = format!(
-        r#"Произведена очистка логов на `{ip}`
-Время выполенения: *{:.2}s*
-Файлов архивировано за период с `{}` по `{}`: *{} файлов*.
-"#,
-        elapsed_time.as_secs_f32(),
-        from_date.format("%d.%m.%Y"),
-        to_date.format("%d.%m.%Y"),
-        amount
-    );
-
-    let res = http
-        .post(&format!("https://api.telegram.org/bot{token}/sendMessage"))
-        .send_json(ureq::json!({
-            "chat_id": chat_id,
-            "text": text,
-            "parse_mode": "markdown"
-        }));
-
-    if let Err(e) = res {
-        let e = e
-            .into_response()
-            .unwrap()
-            .into_json::<ureq::serde_json::Value>()
-            .unwrap();
-
-        eprintln!(
-            "Failed to send message to telegram: {}",
-            e.get("description").unwrap().to_string()
-        );
-    }
-}
-
-fn get_host_ip() -> Option<String> {
-    let mut ip: Option<String> = None;
-
-    for adapter in ipconfig::get_adapters().expect("Failed to get own IP address") {
-        let addr = adapter.ip_addresses()[1].to_string();
-        if addr.starts_with("10.") {
-            ip = Some(addr);
-            break;
-        }
-    }
-
-    return ip;
-}
-
-fn get_http_client(proxy_url: &Option<String>) -> Result<ureq::Agent, ureq::Error> {
-    let mut builder = ureq::AgentBuilder::new();
-
-    if let Some(url) = proxy_url {
-        builder = builder.proxy(ureq::Proxy::new(url)?);
-    }
-
-    Ok(builder.build())
+    println!("Done\n{:.2}s\n{} files", now.elapsed().as_secs_f32(), processed)
 }
 
 fn list_dir_files(path: &PathBuf) -> Vec<DirEntry> {
@@ -187,22 +85,26 @@ fn is_same_day(a: &DateTime<Local>, b: &DateTime<Local>) -> bool {
 fn pack_to_archive(files: &Vec<&DirEntry>, dir: &PathBuf, date: &DateTime<Local>) {
     let human_readable = date.format("%d-%m-%Y");
     let dest = dir.join(format!(
-        "{}_{}.tar",
+        "{}_{}.zip",
         dir.file_name().unwrap().to_str().unwrap(),
         human_readable
     ));
 
     let file = fs::File::create(dest).unwrap();
-    let mut tar = TarBuilder::new(file);
+    // let mut tar = TarBuilder::new(file);
+    let mut zip = zip::ZipWriter::new(file);
+    let options = FileOptions::default()
+        .compression_method(zip::CompressionMethod::DEFLATE);
 
     // Move all files to newly created directory
+    
     for v in files {
-        tar.append_file(v.file_name(), &mut fs::File::open(v.path()).unwrap())
-            .unwrap();
+        zip.start_file(v.file_name().to_string_lossy(), options).unwrap();
+        zip.write_all(&fs::read(v.path()).unwrap()).unwrap();
         fs::remove_file(v.path()).unwrap();
     }
 
-    tar.finish().unwrap();
+    zip.finish().unwrap();
 }
 
 fn archive_files(files: &[DirEntry], parent_dir: &PathBuf) -> usize {
@@ -216,7 +118,7 @@ fn archive_files(files: &[DirEntry], parent_dir: &PathBuf) -> usize {
 
     for f in files.iter() {
         // Skip archives
-        if f.file_name().to_string_lossy().ends_with(".tar") {
+        if f.file_name().to_string_lossy().ends_with(".zip") {
             continue;
         }
 
@@ -256,8 +158,9 @@ fn process_dir(dir: &PathBuf, archive_from: &DateTime<Local>, delete_from: &Date
     });
 
     let mut processed = 0;
+    let len = files.len();
 
-    if files.len() > 0 {
+    if len > 0 {
         // Find start and end indexes for chunk of data that should be proceesed
         // Everything before `end` should be deleted as too old
         let start = files.partition_point(|probe| {
@@ -265,18 +168,21 @@ fn process_dir(dir: &PathBuf, archive_from: &DateTime<Local>, delete_from: &Date
             return time < *delete_from;
         });
 
-        // Everything after start should not be touched as too new
-        let end = start
-            + files[start..].partition_point(|probe| {
-                let time: DateTime<Local> = probe.metadata().unwrap().modified().unwrap().into();
-                return time < *archive_from;
-            });
+        // checking for bounds to not 
+        if start < len {
+            // Everything after start should not be touched as too new
+            let end = start
+                + files[start..].partition_point(|probe| {
+                    let time: DateTime<Local> = probe.metadata().unwrap().modified().unwrap().into();
+                    return time < *archive_from;
+                });
 
-        for deleted in &files[0..start] {
-            fs::remove_file(deleted.path()).unwrap();
+            for deleted in &files[0..start] {
+                fs::remove_file(deleted.path()).unwrap();
+            }
+
+            processed = archive_files(&files[start..end], &dir);
         }
-
-        processed = archive_files(&files[start..end], &dir);
     }
 
     for sub in list_subdirs(&dir) {
